requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
pytz==2024.1
httpx==0.27.0
